requirements.txt
robocorp
robocorp-tasks
robocorp-browser
Flask
PyYAML
flask-cors
selenium
webdriver-manager
requests
python-dotenv
attrs
blinker
certifi
charset-normalizer
click
exceptiongroup
h11
idna
itsdangerous
Jinja2
MarkupSafe
outcome
packaging
PySocks
sniffio
sortedcontainers
trio
trio-websocket
typing_extensions
urllib3
websocket-client
Werkzeug
wsproto
openpyxl